move tests, handle comments
